Fixed real-time neural network issues on Pytorch
