+++ dnn.py
@@ -16,20 +16,21 @@ from copy import deepcopy
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
         
 def initialize_net(model_path):
+
     network = models.resnet50(pretrained=True)
     for param in network.parameters():
        param.requires_grad = False
 
     num_ftrs = network.fc.in_features
     network.fc = nn.Sequential(nn.Linear(num_ftrs, 1), nn.Sigmoid())
-    network = network.to(device)
+    network = nn.DataParallel(network).to(device)
 
     if model_path is not None:
         load_model(model_path, network)
         
     network.eval()
         
-    # Run random sequences to initialize
+    # Run random sequences to initialize bigger chunk of the memory
     placeholder = torch.rand((400, 3, 224, 224)).to(device)
     for i in range(5):
         start = time()
@@ -39,6 +40,7 @@ def initialize_net(model_path):
     return network
 
 def read_stats(file):
+
     with open(file, 'r') as f:
         reader = csv.reader(f, delimiter=",")
         for raw_stats in reader:
@@ -47,41 +49,6 @@ def read_stats(file):
                 "img_std": [float(raw_stats[3]), float(raw_stats[4]), float(raw_stats[5])]
             }
     return stats
-
-class PrepareRTImage(object):
-        
-    def __init__(self, size, stats):
-        self.size = size
-        self.stats = stats
-
-    def normalize(self, image):
-        norm_image = ( np.asarray(image, dtype=float) - np.asarray(self.stats["img_mean"], dtype=float).reshape(1, 1, 3) ) / np.asarray(self.stats["img_std"], dtype=float).reshape(1, 1, 3)
-        return image
-
-    def toTensor(self, data):
-        data = [image.transpose((2, 0, 1)) for image in data] # Channel before x-y
-        return torch.tensor(data, dtype=torch.float)
-
-    def resize(self, image, min_dim):
-        h, w, _ = image.shape
-
-        if h != w: # Erroneous bbox adjustment
-            print("Image dimensions are not the same!")
-        
-        if float(self.size)/float(min(h, w)) != 1:
-            if min(h, w) < self.size:
-                interpolation = cv2.INTER_CUBIC
-            else:
-                interpolation = cv2.INTER_AREA
-        else:
-            return image
-        
-        image = cv2.resize(image, (self.size, self.size), interpolation=interpolation)
-        return image
-    
-    def __call__(self, images):
-        data = [self.normalize(self.resize(image, self.size)) for image in images]        
-        return self.toTensor(data)
 
 def real_time_evaluate(network, data):
 
@@ -95,10 +62,6 @@ def real_time_evaluate(network, data):
 def load_model(model_name, network, optimizer=None):
 
     state = torch.load(model_name)
-    new_state = deepcopy(state)
-    
-    for i in state['network_state'].keys():
-        new_state['network_state'][i.replace('module.', '', 1)] = new_state['network_state'].pop(i)
 
     network.load_state_dict(new_state['network_state'])
     if optimizer is not None:
@@ -106,3 +69,36 @@ def load_model(model_name, network, optimizer=None):
     state = None
     
     print("Model load successful")
+
+class Normalize(object):
+
+    def __init__(self, stats):
+        self.stats = stats
+
+    def __call__(self, image):
+        norm_image = ( np.asarray(image, dtype=float) - np.asarray( self.stats["img_mean"], dtype=float).reshape(1, 1, 3) ) / np.asarray(self.stats["img_std"], dtype=float).reshape(1, 1, 3)
+
+        return norm_image
+
+class Resize(object):
+
+    def __init__(self, size):
+        self.size = size
+
+    def __call__(self, sample):
+        if self.size > sample.shape[0]:
+            interpolation = cv2.INTER_CUBIC
+        else:
+            interpolation = cv2.INTER_AREA
+
+        image = cv2.resize(sample, (self.size, self.size), interpolation=interpolation)
+                
+        return image
+
+class ToTensor(object):
+
+    def __call__(self, sample):
+        image = sample.transpose((2, 0, 1))
+        return torch.tensor(image, dtype=torch.float).unsqueeze(0)
+
+
