+++ state_detect.py
@@ -1,5 +1,6 @@
 import time
 import cv2
+from utils import expand_bbox
 
 def add_padding(rect_par, ratio, dims):
     x, y, w, h = rect_par
@@ -11,27 +12,6 @@ def add_padding(rect_par, ratio, dims):
     else:
         return x-padx, y-pady, w+2*padx, h+2*pady
 
-def expand_bbox(x, y, w, h, width=1920, height=1080):
-    diff = w - h
-
-    if diff > 0: # Wider image. Increase height
-        y -= diff // 2
-        h += diff
-        if y < 0:
-            y = max(0, y)  # Takes care of out of bounds upwards
-        if y + h > height: 
-            y = height - h # Takes care of out of bounds downwards
-        
-    elif diff < 0: # Taller image. Increase width
-        x -= abs(diff) // 2
-        w += abs(diff)
-        if x < 0:
-            x = max(0, x)  # Takes care of out of bounds from left
-        if x + w > width: 
-            x = width - w # Takes care of out of bounds from right
-            
-    return x, y, w, h
-    
 def in_detect_fn(parent):
 
     # Setup
