+++ state_id.py
@@ -1,18 +1,17 @@
 import time
 import cv2
-#import cv2.aruco as aruco
-from dnn import real_time_evaluate, PrepareRTImage, read_stats
+from torchvision.transforms import Compose
+from dnn import real_time_evaluate, PrepareRTImage, read_stats, Normalize, Resize, ToTensor
+from utils import expand_bbox
 
-#aruco_dict = aruco.Dictionary_get(aruco.DICT_4X4_250)
-#parameters = aruco.DetectorParameters_create()
-
-data_prep = PrepareRTImage(224, read_stats("./dataset_stats"))
+data_prep = Compose([ Resize(224), Normalize(read_stats("./dataset_stats")), ToTensor() ])
 
 def in_id_fn(parent):
 
     drone = False
-    predictions = real_time_evaluate(parent.network, data_prep(parent.cur_imgs))
-    #print(predictions)
+    transformed_ims = [data_prep(i) for i in parent.cur_imgs]
+    predictions = real_time_evaluate(parent.network, torch.cat(transformed_ims))
+
     for iter, pred in enumerate(predictions):
         if pred == 1:
             parent.drone_bbox = parent.cur_bboxes[iter]
@@ -23,19 +22,6 @@ def in_id_fn(parent):
         else:
             pass
 
-    '''
-    for i, img in enumerate(parent.cur_imgs):
-        _, ids, _ = aruco.detectMarkers(img, aruco_dict, parameters=parameters)
-
-        if ids is not None and ids[0][0] == 42:
-            parent.drone_bbox = parent.cur_bboxes[i]
-            # print(parent.state)
-            parent.drone()
-            print("drone")
-            return
-        else:
-            pass #print("not_drone")
-    '''
     if not drone:
         parent.not_drone()
 
@@ -49,7 +35,7 @@ def async_id(parent):
         return
 
     vals = [int(a) for a in parent.drone_bbox]
-    x, y, w, h = max(0, vals[0]), max(0, vals[1]), min(parent.camera.width, vals[2]), min(parent.camera.height, vals[3]) 
+    x, y, w, h = expand_bbox(vals)
 
     roi = frame[y:y+h, x:x+w].copy()
     
@@ -57,17 +43,8 @@ def async_id(parent):
     cv2.imshow("async_id", roi)
     cv2.waitKey(1)
 
-    prediction = real_time_evaluate(parent.network, data_prep([roi]))[0]
+    prediction = real_time_evaluate(parent.network, data_prep(roi))[0]
     if prediction == 1:
         return 1
     else:
         return 0
-
-    '''
-    _, ids, _ = aruco.detectMarkers(roi, aruco_dict, parameters=parameters)
-
-    if ids is not None and ids[0][0] == 42:
-        return 1
-    else:
-        return 0
-    '''
